Add snowpack as part of build process
Ensures that we can use dependencies from npm. Is not really a
packer/bundler - simply copies dependencies to a public directory and
rewrites imports to serve correctly, allowing for native JS execution.
This is bounds better than anything else, as no bundler code is actually
running on the client; it's just a tool for me.