Implement client-side FSM mechanism
Client maintains a state interface (which is simply a JS version of a
function pointer for our use case). This function is called when a
message is received and returns a new state. This allows us to shift
state based on what messages come in, which lets the client be entirely
predictable in response to events.