Implement message loop and PING system
Player runner is now split into two halves: the read half and the write
half. Both are controlled by the same context, which will cause both to
exit on cancellation/timeout. The message loop will simply continually
read messages from the user, but will timeout after the expected time
within which a PING could arrive. This means that the server will expect
keep-alive PINGs from the client within this time or the client will be
considered to have lost connection (20s is the timeout). The PING system
will send a ping control packet to the client every 15 seconds. It will
not wait for a response: the read half will obtain the response.