Split client logic from specific client type
Created a new, generic "Client" struct, consisting of an io.ReadCloser
compatible interface and a writer half for the websocket, allowing us to
have the same PING system functionality as before. This can then be used
or embedded into further client types (such as player or host), which
eliminates code duplication for websocket handling.

TODO: Move connection handling over to using a temporary Client object
for the transaction.