Remove test game generation
We are so close to adding actual game creation that I feel it is not
very uesful to keep creating one. I may add a proper mechanism for
creating one in debug builds later on, which could laod the default quiz
or one from memory. For now, this will encourage me to keep working on
host UX rather than keep tweaking user UX.