Implement host player roster management
Host will now remove players which it knows have been kicked. Before the
server responds to the kick request, the player is in the "loading"
state, which the host is aware of and can show visual feedback.

Host will also now respond to disconnects, showing a red, crossed
username.

TODO: Show visual feedback on loading state