Fix possible game runner deadlock
If client or host disconnects while attempting to send a message, the
calling thread will block indefinitely, causing a deadlock. So far, this
has not been noticeable, as communication was passed in parallel for
performance, but this simply deferred the problem to a leaked thread
rather than a deadlock.

Now, if the communicating writer half has been shut down by a context
call, the write call will error immediately, preventing deadlock.