Begin to implement question flow
Host orders the next question via the "next" command to the server. This
causes a state shift and some new states in the FSM which handle
invariance of question state.

Answers are only accepted from the players after a message is received
from the host which states that the question countdown is over. This
means that we rely entirely upon the host's timing, which is much more
reliable than trying to coordinate timings from server to clients
reliably. This does, however, mean that the host can ruin their own
game by cheating, for instance.

TODO: Frontend for question flow
TODO: Answer submissions, checking and feedback