Improve reliability of ``pack`` script
Reliably detect import statements, rather than simply scanning for any
line which appears to contain a from declaration. This eliminates any
lines which are comments or simply refer to identifiers containing
"from". For instance, the identifier "variable_from" would be
incorrectly detected as an import. This should cover all the corner
cases now.